sprite redraw experiment
WIP…

experiments in sprite handling

drawing code includes an `updateRect` which tracks the rectangular area of the screen that will be affected by draw operations when processing the drawing queue.  this value was never actually used

previously `hideSprites` would just blindly hide _all_ sprites, and `showSprites` similarly just show them all.

this commit changes `hideSprites` to only hide sprites that are overlap with the `updateRect` (and those sprites that also overlap with them).  `showSprites` will then only show those sprites that are tracked as being currently hidden

the idea of this change is to potentially reduce the numbers of sprites that need to be hidden